Conflicts:
	jni/gles2rice/src/RenderBase.cpp

+++ jni/gles2rice/src/RSP_Parser.cpp
@@ -1303,7 +1303,7 @@ void DLParser_FillRect(Gfx *gfx)
             // Emulating Clear, by write the memory in RDRAM
             uint16 color = (uint16)gRDP.originalFillColor;
             uint32 pitch = g_CI.dwWidth<<1;
-            long base = (long) (g_pRDRAMu8 + g_CI.dwAddr);
+            long long base = (long long) (g_pRDRAMu8 + g_CI.dwAddr);
             for( uint32 i =y0; i<y1; i++ )
             {
                 for( uint32 j=x0; j<x1; j++ )
@@ -1330,7 +1330,7 @@ void DLParser_FillRect(Gfx *gfx)
             {
                 uint16 color = (uint16)gRDP.originalFillColor;
                 uint32 pitch = g_pRenderTextureInfo->N64Width<<1;
-                long base = (long) (g_pRDRAMu8 + g_pRenderTextureInfo->CI_Info.dwAddr);
+                long long base = (long long) (g_pRDRAMu8 + g_pRenderTextureInfo->CI_Info.dwAddr);
                 for( uint32 i =y0; i<y1; i++ )
                 {
                     for( uint32 j=x0; j<x1; j++ )
@@ -1343,7 +1343,7 @@ void DLParser_FillRect(Gfx *gfx)
             {
                 uint8 color = (uint8)gRDP.originalFillColor;
                 uint32 pitch = g_pRenderTextureInfo->N64Width;
-                long base = (long) (g_pRDRAMu8 + g_pRenderTextureInfo->CI_Info.dwAddr);
+                long long base = (long long) (g_pRDRAMu8 + g_pRenderTextureInfo->CI_Info.dwAddr);
                 for( uint32 i=y0; i<y1; i++ )
                 {
                     for( uint32 j=x0; j<x1; j++ )
@@ -1420,7 +1420,7 @@ void DLParser_FillRect(Gfx *gfx)
         DebuggerAppendMsg("Pause after FillRect: Color=%08X\n", gRDP.originalFillColor);});
     }
 }
-//Nintro64 uses Sprite2d 
+//Nintro64 uses Sprite2d
 void RSP_RDP_Nothing(Gfx *gfx)
 {
     SP_Timing(RSP_RDP_Nothing);
@@ -1444,10 +1444,10 @@ void RSP_RDP_Nothing(Gfx *gfx)
         return;
     }
 #endif
-        
+
     if( options.bEnableHacks )
         return;
-    
+
     gDlistStackPointer=-1;
 }
 
@@ -1497,7 +1497,7 @@ void RSP_RDP_InsertMatrix(Gfx *gfx)
     }
     else
     {
-        if( pauseAtNext && logMatrix ) 
+        if( pauseAtNext && logMatrix )
         {
             DebuggerAppendMsg("insert matrix: %08X, %08X", gfx->words.cmd0, gfx->words.cmd1);
         }
